Add test that verifies a command is supplied to
- In a refactor from a previous commit, we moved where we verify that a
command is supplied. One of our tests ended up accidentally testing this
case when the intention was to test that exactly 4 arguments were given.
Therefore, I added a test for requiring a command and adjusted the test
that checked for 4 arguments.
- This commit also includes a more complex regexp that is used in the
mysql version test. The version can now be in the format of
`x.x.x-build.x (x)` as well as `x.x.x (x)`

[#156061064]