Remove unnecessary WaitForService step in acceptance test
The plugin now attempts to update a service and enable tls as part of
the migration process.  This means the service instance will never be in
a "create succeeded" state but rather either "update succeeded" or
"update failed".  If the plugin exits successfully we should assume the
service instance is in a usable state or allow the subsequent steps to
just fail.

[#156453821]