Stop failing hostname resolution fast for non temporary DNS errors
The temporary flag is only set for network timeout related DNS errors.
It doesn't seem sufficient for our current use case. If we did get this
wrong, our assumption is that in the worst case, we have added 1 minute
to failing migrations.

This is to fix the increased number of flaky builds that we are seeing since PCF 2.8 was
released. We were seeing the builds only wait 1 second for DNS
resolution, which was clearly not enough.

[#170432284]

Signed-off-by: Jamie Monserrate <jmonserrate@pivotal.io>