Remove test assertion around TLS in service keys
In old mysql tiles, enable_tls needed to be run by the developer
manually in order to enable TLS on the service instance.  In modern
MySQL Tiles (>= MySQL 2.5) TLS is automatically enabled if configured by
the operator.

This test assertion is no longer useful and prevents the acceptance test
suite from being run in an environment with TLS disabled.

[#171027755](https://www.pivotaltracker.com/story/show/171027755)

Signed-off-by: Andrew Garner <agarner@pivotal.io>