run docker containers via docker cli
This removes a dependency on various Go libraries required to interact
with docker directly.  These dependencies tend to have vulnerabilities
and need to be frequently updated.

Since docker is only used for testing, all test suites that interact
with docker have been updated to user the docker cli rather than a Go
library significantly reducing the set of tracked dependencies and
simplifying testing.