extract build dependencies into a discrete Dockerfile
This avoids dynamically installing various dependencies in a bash script
and instead rely on a docker image.

In CI, the idea is that this image would be built and used for any CI
jobs that build the plugin.

Locally, the build-plugin script will build this image on demand and use
it before building the assets required by the plugin.

As part of this, added additional caching of the build dependencies for
local development.  The build-plugin script now mounts cache volumes for
/go and /build directories to cache Go compilation and Percona Server
compilation (respectively).