inject app.Extractor dependency into plugin to avoid tight coupling
The app extractor is only usable when the migration app is built
otherwise go:embed will not find the assets and go will refuse to
compile that package.

If assets are not built prior to running unit tests, any _other_ package
coupled with the "app" package will fail.

Now, only "main" imports this package and passes it as a dependency to
the MySQLPlugin for use in the migrate job, rather than the plugin
directly depending on the app package. This dependency injection
improves testability of the plugin, which has many historical unit
testing gaps and allows us to improve this testing in the future.