specs: allow up to 5 minutes for app to start
Previously, this helper implicitly waited for 4m due to the
implementation of ExecuteCfCmd. When implementing reattempts the
timeout was changed to 60 seconds incorrectly.  Spring apps typically
take a couple minutes to start and 60s is too low.

Now bump up the timeout to a "round" 5 minutes to wait for an app to start.

Authored-by: Andrew Garner <garnera@vmware.com>