multi-site targets prefer CF_PLUGIN_HOME
This behavior was reverted during an earlier refactor but in hindsight
this was a mistake.

The targeting helpers will now use one of the following directories:

- $CF_PLUGIN_HOME/.cf/.mysql-tools/
- $CF_HOME/.cf/.mysql-tools/
- $HOME/.cf/.mysql-tools/

This depends on whether the environment variables CF_PLUGIN_HOME or
CF_HOME are set.

When saving a configuration, the cf cli will use one of:

- $CF_HOME/.cf/confgi.json
- $HOME/.cf/config.json

The save-target command will use the path advertised by the internal cf
cli package `confighelpers.DefaultFilePath()`.

This purpose of this change is to allow a user to specify a customer
CF_HOME to allow targetting different foundations in different terminal
sessions, but having a single, shared CF_PLUGIN_HOME where targets can
be saved in order to make it easy to target multiple environments.

[#186582809](https://www.pivotaltracker.com/story/show/186582809)