Adds new switchover command
Creates new workflow for the switchover process

This process treats HA and single-node plans the same. It will
perform a switchover based on checking the plan types of the leader and
the follower. The switchover will use those plans when setting up the
new leader and follower.

The end result is that the "new leader" will have the same plan type as
the "old leader", and same thing for the follower.

Co-authored-by: Kyle Ong <kyleo@vmware.com>
Co-authored-by: Kim Bassett <kbassett@vmware.com>
Co-authored-by: Ryan Wittrup <rwittrup@vmware.com>